refactor: consolidate type.h by including bridge.h and removing redundant code
